fix 64-bit build
git-svn-id: svn+ssh://svn.freebsd.org/base/head@186341 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.sbin/makefs/ffs.c
@@ -371,7 +371,7 @@ ffs_validate(const char *dir, fsnode *root, fsinfo_t *fsopts)
 			    "maxsize of %lld; rounding down to %lld.",
 			    __func__, dir, (long long)size,
 			    (long long)fsopts->maxsize,
-			    rounddown(fsopts->size, fsopts->bsize));
+			    (long long) rounddown(fsopts->size, fsopts->bsize));
 		}
 		size = rounddown(fsopts->size, fsopts->bsize);
 	}
